Simplify project

+++ test/packaging/find_package/hello.cc
@@ -4,6 +4,6 @@
 #include <iostream> // std::cout
 
 auto main() -> int {
-  std::cout << sourcemeta::hydra::http::Status::OK << "\n";
+  std::cout << static_cast<int>(sourcemeta::hydra::http::Status::OK) << "\n";
   return EXIT_SUCCESS;
 }
